# server/openapi.json
{
  "components": {
    "schemas": {
...
      "CodeBuddyLoginResponse": {
        "properties": {
          "authorize_url": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "authorize_url",
          "state"
        ],
        "title": "CodeBuddyLoginResponse",
        "type": "object"
      },
...
      "ProviderConnectionView": {
        "description": "A connection as the detail response reports it. There is deliberately no\ncredential field: a stored API key never leaves the database.",
        "properties": {
          "alias": {
            "type": [
              "string",
              "null"
            ]
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": "string"
          },
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "provider_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_id",
          "name",
          "category",
          "protocol",
          "enabled",
          "created_at"
        ],
        "type": "object"
      },
      "ProviderEntry": {
        "description": "A provider as the list, catalog, and detail routes describe it.",
        "properties": {
          "category": {
            "type": "string"
          },
          "connections": {
            "description": "Detail only; the list and catalog omit it like Node does.",
            "items": {
              "$ref": "#/components/schemas/ProviderConnectionView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "default_base_url": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ProviderModel"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/ProviderProtocol"
          },
          "requires_api_key": {
            "type": "boolean"
          },
          "requires_oauth": {
            "type": "boolean"
          },
          "round_robin": {
            "description": "Whether requests rotate across this provider's accounts.",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/ProviderStatus"
          },
          "supports_custom_url": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "protocol",
          "default_base_url",
          "requires_api_key",
          "requires_oauth",
          "supports_custom_url",
          "enabled",
          "round_robin",
          "status",
          "models"
        ],
        "type": "object"
      },
...
      "ProviderModel": {
        "description": "A model line in the detail response. Hidden and favorited models stay in the\nlist so the admin view can restore them; `GET /v1/models` is the route that\ndrops them.",
        "properties": {
          "favorite": {
            "type": "boolean"
          },
          "hidden": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "owned_by": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "object",
          "owned_by",
          "hidden",
          "favorite"
        ],
        "type": "object"
      },
      "ProviderProtocol": {
        "description": "The wire protocol a provider speaks, as the catalog and the provider-auth\nresponses report it.\n\nThree variants, matching what the build actually serves. Node's\n`ProviderProtocol` union also lists `gemini`, but that value is dead: its only\nuser was the `gemini_cli` provider deleted with `packages/providers/src/catalog.ts`.",
        "enum": [
          "openai",
          "anthropic",
          "custom"
        ],
        "type": "string"
      },
...
      "ProviderStatus": {
        "description": "Connection state of a provider driver.",
        "properties": {
          "connected_count": {
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "state",
          "message",
          "connected_count"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminSession": {
        "description": "The admin session cookie. It also satisfies API-key auth on the routes that accept it.",
        "in": "cookie",
        "name": "srouter_admin_session",
        "type": "apiKey"
      },
      "bearerAuth": {
        "description": "The same API key in the `Authorization` header.",
        "scheme": "bearer",
        "type": "http"
      },
      "srouterApiKey": {
        "description": "An SRouter API key. `Authorization: Bearer <key>` carries the same credential.",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Multi-provider OpenAI and Anthropic compatible LLM gateway, as served by the Rust build (`server/`). One listener on `PORT` carries the API, the dashboard SPA, and the provider callbacks; there is no secondary OAuth listener. Every response uses snake_case field names (owner ruling 2026-10-07), and errors use the `{error:{message,type,code?,param?}}` envelope. Cloudflare Tunnel routes and `/v1/settings/fallbacks/*` are excluded from this build by owner ruling; they do not appear below and answer `404`.",
    "title": "SRouter API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/codebuddy-cn/login": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeBuddyLoginResponse"
                }
              }
            },
            "description": "With `format=json`: the authorization URL."
          },
          "302": {
            "description": "Without `format=json`: a redirect to the authorization URL."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Start the CodeBuddy device flow"
      }
    },
    "/v1/auth/codebuddy-cn/poll": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "The flow result."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Poll the CodeBuddy device flow"
      },
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "The flow result."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Poll the CodeBuddy device flow"
      }
    },
    "/v1/auth/codebuddy/login": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeBuddyLoginResponse"
                }
              }
            },
            "description": "With `format=json`: the authorization URL."
          },
          "302": {
            "description": "Without `format=json`: a redirect to the authorization URL."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Start the CodeBuddy device flow"
      }
    },
    "/v1/auth/codebuddy/poll": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "The flow result."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Poll the CodeBuddy device flow"
      },
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "The flow result."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Poll the CodeBuddy device flow"
      }
    },
...
    "/v1/providers/{provider_id}": {
      "get": {
        "parameters": [
          {
            "description": "Provider id, for example `qoder`.",
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEntry"
                }
              }
            },
            "description": "The provider detail entry, with its connections."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "Read one provider"
      },
      "patch": {
        "parameters": [
          {
            "description": "Provider id, for example `qoder`.",
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEntry"
                }
              }
            },
            "description": "The detail entry read back after the write."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Toggle a provider or its model flags"
      }
    },
...
  }
}